Complete DB processing & Refactor

+++ sql/create_tables/10.sql
@@ -6,7 +6,7 @@ CREATE TABLE header_10(
     custodian_name TEXT,
     custodian_code INTEGER,
     process_date DATE,
-    volume_number INTEGER,
+    volume_number SMALLINT,
     entry_date DATE,
     time_stamp TIME WITHOUT TIME ZONE,
     dtf_version NUMERIC(3,2),
